requirements.txt
sqlalchemy
pymysql
apache-airflow==1.10.12
Faker==4.1.3
cryptography==3.1
onetimepad==1.4
pycrypto
pandas_profiling==2.9.0
